+++ cmd/hidden_lake/service/internal/handler/network_request.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"net/http"
+	"sync"
 
 	"github.com/number571/go-peer/cmd/hidden_lake/service/internal/config"
 	"github.com/number571/go-peer/cmd/hidden_lake/service/pkg/request"
@@ -25,7 +26,7 @@ const (
 	cErrorLoadRequestID
 )
 
-func HandleNetworkRequestAPI(pCtx context.Context, pConfig config.IConfig, pLogger logger.ILogger, pNode anonymity.INode) http.HandlerFunc {
+func HandleNetworkRequestAPI(pCtx context.Context, pMutex *sync.Mutex, pConfig config.IConfig, pLogger logger.ILogger, pNode anonymity.INode) http.HandlerFunc {
 	return func(pW http.ResponseWriter, pR *http.Request) {
 		logBuilder := http_logger.NewLogBuilder(pkg_settings.CServiceName, pR)
 
@@ -67,7 +68,7 @@ func HandleNetworkRequestAPI(pCtx context.Context, pConfig config.IConfig, pLogg
 			panic("undefined error code")
 		}
 
-		if ok, err := setRequestID(pNode, requestID); err != nil {
+		if ok, err := setRequestID(pMutex, pNode, requestID); err != nil {
 			if ok {
 				pLogger.PushWarn(logBuilder.WithMessage("request_id_exist"))
 				api.Response(pW, http.StatusBadGateway, "failed: request id exist")