+++ cmd/hidden_lake/service/pkg/app/app.go
@@ -30,9 +30,10 @@ var (
 )
 
 type sApp struct {
-	fState state.IState
+	fState  state.IState
+	fPathTo string
 
-	fPathTo     string
+	fMutex      *sync.Mutex
 	fCfgW       config.IWrapper
 	fNode       anonymity.INode
 	fConnKeeper conn_keeper.IConnKeeper
@@ -52,6 +53,7 @@ func NewApp(
 	pPathTo string,
 	pParallel uint64,
 ) types.IRunner {
+	mtx := &sync.Mutex{}
 	logging := pCfg.GetLogging()
 
 	var (
@@ -61,11 +63,12 @@ func NewApp(
 	)
 
 	cfgWrapper := config.NewWrapper(pCfg)
-	node := initNode(cfgWrapper, pPrivKey, anonLogger, pParallel)
+	node := initNode(mtx, cfgWrapper, pPrivKey, anonLogger, pParallel)
 
 	return &sApp{
 		fState:      state.NewBoolState(),
 		fPathTo:     pPathTo,
+		fMutex:      mtx,
 		fCfgW:       cfgWrapper,
 		fNode:       node,
 		fConnKeeper: initConnKeeper(pCfg, node),