+++ cmd/hidden_lake/service/pkg/app/init_node.go
@@ -1,6 +1,7 @@
 package app
 
 import (
+	"sync"
 	"time"
 
 	"github.com/number571/go-peer/cmd/hidden_lake/service/internal/config"
@@ -19,7 +20,7 @@ import (
 	"github.com/number571/go-peer/pkg/client"
 )
 
-func initNode(pCfgW config.IWrapper, pPrivKey asymmetric.IPrivKey, pLogger logger.ILogger, pParallel uint64) anonymity.INode {
+func initNode(pMutex *sync.Mutex, pCfgW config.IWrapper, pPrivKey asymmetric.IPrivKey, pLogger logger.ILogger, pParallel uint64) anonymity.INode {
 	cfg := pCfgW.GetConfig()
 	cfgSettings := cfg.GetSettings()
 	queueDuration := time.Duration(cfgSettings.GetQueuePeriodMS()) * time.Millisecond
@@ -87,6 +88,6 @@ func initNode(pCfgW config.IWrapper, pPrivKey asymmetric.IPrivKey, pLogger logge
 		}(),
 	).HandleFunc(
 		pkg_settings.CServiceMask,
-		handler.HandleServiceTCP(pCfgW, pLogger),
+		handler.HandleServiceTCP(pMutex, pCfgW, pLogger),
 	)
 }