+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Update `cmd/hidden_lake/_template`: _template -> helpers/template
 - Update `cmd/hidden_lake/composite`: composite -> composites
 - Update `Makefile`: delete from test-coverage examples, tools, micro_anon, cmd directories
+- Update `cmd/hidden_lake/service`: delete global mutex from handler functions
 
 ### BUG FIXES